+++ core/const.py
@@ -1,6 +1,6 @@
 """Constants used by Photos.network core."""
 MAJOR_VERSION = 0
-MINOR_VERSION = 3
+MINOR_VERSION = 5
 PATCH_VERSION = 0
 __short_version__ = f"{MAJOR_VERSION}.{MINOR_VERSION}"
 __version__ = f"{__short_version__}.{PATCH_VERSION}"